+++ src/constants.js
@@ -1,11 +1,14 @@
 import url from 'url';
 
+export const DATA_FORMAT_VERSION = '1.2';
+
 export const HOOKS = [
     'before-invocation',
     'after-invocation',
 ];
 
 export const URL = url.parse(
+    // the comment below is for ignoring in unit tests, do not remove it
     // istanbul ignore next
     process.env.thundra_lambda_publish_rest_baseUrl
         ? process.env.thundra_lambda_publish_rest_baseUrl